+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.0] - 2023-11-21
+### Added
+- Adds config to add BurgerIcon to the TabBar.
+- Adds configuration to add content to the Category Drawer.
+- Adds compatibility with extension [@shopgate-project/page-switcher](https://github.com/shopgate-professional-services/ext-page-switcher)
+
 ## [1.0.0] - 2022-07-27
 ### Added
 - Adds BurgerIcon to open a Category Drawer.